Rename bin from sandbox-ce to sourced
